// frontend/src/types/ai.ts
export interface TiterInputParams {
  temperature: number;
  substrate_concentration: number;
  reactor_volume: number;
  oxygen: number;
  fermentation_duration: number;
}